refactor(web): controller 패키지로 분리
