+++ src/main/java/roomescape/web/controller/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.web;
+package roomescape.web.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.core.domain.ReservationTime;
-import roomescape.web.request.ReservationTimeRequest;
-import roomescape.web.response.ReservationTimeResponse;
+import roomescape.web.controller.request.ReservationTimeRequest;
+import roomescape.web.controller.response.ReservationTimeResponse;
 import roomescape.core.repository.ReservationTimeRepository;
 
 import java.util.List;
