+++ src/main/java/roomescape/web/controller/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.web.request;
+package roomescape.web.controller.request;
 
 import roomescape.core.domain.ReservationTime;
