+++ src/main/java/roomescape/web/controller/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.response;
+package roomescape.web.controller.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import roomescape.core.domain.ReservationTime;
